requirements.txt
python-pptx
Pillow
playwright

